// src/app/Model/Customer.ts
export interface Car {
    id: string;
    bname: string;
    mname: string;
    price: string;
    condition: string;
    status: string
}
export interface Customer {
    id: string;
    name: string;
    email: string;
    phone: string;
    active: string
}
export interface Country{
    code:string,
    name:string
}